add v2.1.4

+++ Examples/I2C/Fast_mode_ plus_communication/main.c
@@ -11,27 +11,27 @@
 /*
     Copyright (c) 2020, GigaDevice Semiconductor Inc.
 
-    Redistribution and use in source and binary forms, with or without modification, 
+    Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
 
-    1. Redistributions of source code must retain the above copyright notice, this 
+    1. Redistributions of source code must retain the above copyright notice, this
        list of conditions and the following disclaimer.
-    2. Redistributions in binary form must reproduce the above copyright notice, 
-       this list of conditions and the following disclaimer in the documentation 
+    2. Redistributions in binary form must reproduce the above copyright notice,
+       this list of conditions and the following disclaimer in the documentation
        and/or other materials provided with the distribution.
-    3. Neither the name of the copyright holder nor the names of its contributors 
-       may be used to endorse or promote products derived from this software without 
+    3. Neither the name of the copyright holder nor the names of its contributors
+       may be used to endorse or promote products derived from this software without
        specific prior written permission.
 
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
-IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
-INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT 
-NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR 
-PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
-WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
+    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
 OF SUCH DAMAGE.
 */
 
@@ -49,7 +49,7 @@ __IO ErrStatus state = ERROR;
 void rcu_config(void);
 void gpio_config(void);
 void i2c_config(void);
-ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length);
+ErrStatus memory_compare(uint8_t *src, uint8_t *dst, uint16_t length);
 
 /*!
     \brief      main function
@@ -59,19 +59,19 @@ ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length);
 */
 int main(void)
 {
-    int i;
-    
+    uint8_t i;
+
     /* initialize led */
     gd_eval_led_init(LED2);
     gd_eval_led_init(LED3);
-    /* RCU config */
+    /* configure RCU */
     rcu_config();
-    /* GPIO config */
+    /* configure GPIO */
     gpio_config();
-    /* I2C config */
+    /* configure I2C */
     i2c_config();
-    
-    for(i=0; i<16; i++){
+
+    for(i = 0; i < 16; i++) {
         i2c_transmitter[i] = i + 0x80;
     }
     /* wait until I2C bus is idle */
@@ -90,8 +90,8 @@ int main(void)
     while(!i2c_flag_get(I2C1, I2C_FLAG_ADDSEND));
     /* clear ADDSEND bit */
     i2c_flag_clear(I2C1, I2C_FLAG_ADDSEND);
-    
-    for(i=0; i<15; i++){
+
+    for(i = 0; i < 15; i++) {
         /* send a data byte */
         i2c_data_transmit(I2C1, i2c_transmitter[i]);
         /* wait until the transmission data register is empty */
@@ -111,24 +111,24 @@ int main(void)
     while(!i2c_flag_get(I2C1, I2C_FLAG_AERR));
     /* send a stop condition to I2C bus */
     i2c_stop_on_bus(I2C0);
-    while(I2C_CTL0(I2C0)&0x0200);
+    while(I2C_CTL0(I2C0) & 0x0200);
     i2c_receiver[i] = i2c_data_receive(I2C0);
     i2c_ack_config(I2C0, I2C_ACK_ENABLE);
     /* clear the bit of AE */
     i2c_flag_clear(I2C1, I2C_FLAG_AERR);
-    
+
     /* compare the transmit buffer and the receive buffer */
     state = memory_compare(i2c_transmitter, i2c_receiver, 16);
     /* if success, LED2 and LED3 are on */
-    if(SUCCESS == state){
+    if(SUCCESS == state) {
         gd_eval_led_on(LED2);
         gd_eval_led_on(LED3);
-    }else{
+    } else {
         /* if failed, LED2 and LED3 are off */
         gd_eval_led_off(LED2);
-        gd_eval_led_off(LED3); 
+        gd_eval_led_off(LED3);
     }
-    while(1){
+    while(1) {
     }
 }
 
@@ -140,10 +140,10 @@ int main(void)
     \param[out] none
     \retval     ErrStatus : ERROR or SUCCESS
 */
-ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length) 
+ErrStatus memory_compare(uint8_t *src, uint8_t *dst, uint16_t length)
 {
-    while(length--){
-        if(*src++ != *dst++){
+    while(length--) {
+        if(*src++ != *dst++) {
             return ERROR;
         }
     }
