+++ Examples/I2C/Master_receiver&slave_transmitter_interrupt/main.c
@@ -11,27 +11,27 @@
 /*
     Copyright (c) 2020, GigaDevice Semiconductor Inc.
 
-    Redistribution and use in source and binary forms, with or without modification, 
+    Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
 
-    1. Redistributions of source code must retain the above copyright notice, this 
+    1. Redistributions of source code must retain the above copyright notice, this
        list of conditions and the following disclaimer.
-    2. Redistributions in binary form must reproduce the above copyright notice, 
-       this list of conditions and the following disclaimer in the documentation 
+    2. Redistributions in binary form must reproduce the above copyright notice,
+       this list of conditions and the following disclaimer in the documentation
        and/or other materials provided with the distribution.
-    3. Neither the name of the copyright holder nor the names of its contributors 
-       may be used to endorse or promote products derived from this software without 
+    3. Neither the name of the copyright holder nor the names of its contributors
+       may be used to endorse or promote products derived from this software without
        specific prior written permission.
 
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
-IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
-INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT 
-NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR 
-PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
-WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
+    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
 OF SUCH DAMAGE.
 */
 
@@ -42,8 +42,8 @@ OF SUCH DAMAGE.
 uint8_t i2c_buffer_transmitter[16];
 uint8_t i2c_buffer_receiver[16];
 
-volatile uint8_t*       i2c_txbuffer;
-volatile uint8_t*       i2c_rxbuffer;
+volatile uint8_t       *i2c_txbuffer;
+volatile uint8_t       *i2c_rxbuffer;
 volatile uint16_t       I2C_nBytes;
 volatile ErrStatus      status;
 ErrStatus state = ERROR;
@@ -52,7 +52,7 @@ void rcu_config(void);
 void gpio_config(void);
 void i2c_config(void);
 void i2c_nvic_config(void);
-ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length);
+ErrStatus memory_compare(uint8_t *src, uint8_t *dst, uint16_t length);
 
 /*!
     \brief      main function
@@ -62,8 +62,8 @@ ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length);
 */
 int main(void)
 {
-    int i;
-    
+    uint8_t i;
+
     /* initialize LED2, LED3, as the transfer instruction */
     gd_eval_led_init(LED2);
     gd_eval_led_init(LED3);
@@ -72,15 +72,15 @@ int main(void)
     i2c_config();
     i2c_nvic_config();
 
-    for(i=0; i<16; i++){
+    for(i = 0; i < 16; i++) {
         i2c_buffer_transmitter[i] = i + 0x80;
     }
-     /* initialize i2c_txbuffer, i2c_rxbuffer, I2C_nBytes and status */
+    /* initialize i2c_txbuffer, i2c_rxbuffer, I2C_nBytes and status */
     i2c_txbuffer = i2c_buffer_transmitter;
     i2c_rxbuffer = i2c_buffer_receiver;
     I2C_nBytes = 16;
     status = ERROR;
-    
+
     /* enable the I2C0 interrupt */
     i2c_interrupt_enable(I2C0, I2C_INT_ERR);
     i2c_interrupt_enable(I2C0, I2C_INT_BUF);
@@ -89,8 +89,8 @@ int main(void)
     i2c_interrupt_enable(I2C1, I2C_INT_ERR);
     i2c_interrupt_enable(I2C1, I2C_INT_BUF);
     i2c_interrupt_enable(I2C1, I2C_INT_EV);
-    
-    if(2 == I2C_nBytes){
+
+    if(2 == I2C_nBytes) {
         /* send ACK for the next byte */
         i2c_ackpos_config(I2C0, I2C_ACKPOS_NEXT);
     }
@@ -98,20 +98,20 @@ int main(void)
     while(i2c_flag_get(I2C0, I2C_FLAG_I2CBSY));
     /* the master sends a start condition to I2C bus */
     i2c_start_on_bus(I2C0);
-    
-    while(I2C_nBytes>0);
+
+    while(I2C_nBytes > 0);
     while(SUCCESS != status);
     /* if the transfer is successfully completed, LED2 and LED3 is on */
     state = memory_compare(i2c_buffer_transmitter, i2c_buffer_receiver, 16);
-    if(SUCCESS == state){
+    if(SUCCESS == state) {
         /* if success, LED2 and LED3 are on */
         gd_eval_led_on(LED2);
         gd_eval_led_on(LED3);
-    }else{
+    } else {
         /* if failed, LED2 and LED3 are off */
         gd_eval_led_off(LED2);
-        gd_eval_led_off(LED3); 
-    } 
+        gd_eval_led_off(LED3);
+    }
     while(1)
     {}
 }
@@ -124,10 +124,10 @@ int main(void)
     \param[out] none
     \retval     ErrStatus : ERROR or SUCCESS
 */
-ErrStatus memory_compare(uint8_t* src, uint8_t* dst, uint16_t length) 
+ErrStatus memory_compare(uint8_t *src, uint8_t *dst, uint16_t length)
 {
-    while(length--){
-        if(*src++ != *dst++){
+    while(length--) {
+        if(*src++ != *dst++) {
             return ERROR;
         }
     }
@@ -173,9 +173,9 @@ void gpio_config(void)
 */
 void i2c_config(void)
 {
-    /* I2C clock configure */
+    /* configure I2C clock */
     i2c_clock_config(I2C0, 100000, I2C_DTCY_2);
-    /* I2C address configure */
+    /* configure I2C address */
     i2c_mode_addr_config(I2C0, I2C_I2CMODE_ENABLE, I2C_ADDFORMAT_7BITS, I2C0_SLAVE_ADDRESS7);
     /* enable I2C0 */
     i2c_enable(I2C0);
